Move multi-driven tests to analysis layer
