+++ tests/unittests/ast/MemberTests.cpp
@@ -1310,49 +1310,6 @@ endmodule
     CHECK(diags[1].code == diag::BlockingInAlwaysFF);
 }
 
-TEST_CASE("always_comb drivers within nested functions") {
-    auto tree = SyntaxTree::fromText(R"(
-module m;
-    int baz;
-
-    function void f1(int bar);
-      baz = bar;
-    endfunction
-
-    function void f2(int bar);
-      f1(bar);
-    endfunction
-
-    always_comb f2(2);
-    always_comb f2(3);
-
-    int v;
-    function void f3(int bar);
-      v = bar;
-    endfunction
-
-    always_comb f3(4);
-
-    int foo;
-    task t;
-      foo <= 1;
-    endtask
-
-    always_comb begin
-      foo <= 2;
-      t();
-    end
-endmodule
-)");
-
-    Compilation compilation;
-    compilation.addSyntaxTree(tree);
-
-    auto& diags = compilation.getAllDiagnostics();
-    REQUIRE(diags.size() == 1);
-    CHECK(diags[0].code == diag::MultipleAlwaysAssigns);
-}
-
 TEST_CASE("always_comb timing inside assertion") {
     auto tree = SyntaxTree::fromText(R"(
 module m;
@@ -1434,28 +1391,6 @@ endmodule
     NO_COMPILATION_ERRORS;
 }
 
-TEST_CASE("always_comb dup driver with initial block with language option") {
-    auto tree = SyntaxTree::fromText(R"(
-module m;
-    int foo[2];
-
-    initial begin
-        for (int i = 0; i < 2; i++)
-            foo[i] = 0;
-    end
-
-    always_comb foo[1] = 1;
-endmodule
-)");
-
-    CompilationOptions options;
-    options.flags |= CompilationFlags::AllowDupInitialDrivers;
-
-    Compilation compilation(options);
-    compilation.addSyntaxTree(tree);
-    NO_COMPILATION_ERRORS;
-}
-
 TEST_CASE("always_ff timing control restrictions") {
     auto tree = SyntaxTree::fromText(R"(
 module m;
@@ -1490,72 +1425,6 @@ endmodule
     CHECK(diags[0].code == diag::AlwaysFFEventControl);
     CHECK(diags[1].code == diag::AlwaysFFEventControl);
     CHECK(diags[2].code == diag::BlockingInAlwaysFF);
-}
-
-TEST_CASE("hierarchical driver errors") {
-    auto tree = SyntaxTree::fromText(R"(
-interface I;
-    int foo;
-endinterface
-
-module m;
-    I i();
-
-    n n1(i);
-    n n2(i);
-endmodule
-
-module n(I i);
-    always_comb i.foo = 1;
-endmodule
-)");
-
-    Compilation compilation;
-    compilation.addSyntaxTree(tree);
-
-    auto& diags = compilation.getAllDiagnostics();
-    std::string result = "\n" + report(diags);
-    CHECK(result == R"(
-source:14:17: error: variable 'foo' driven by always_comb procedure cannot be written to by any other process
-    always_comb i.foo = 1;
-                ^~~~~
-note: from 'm.n2' and 'm.n1'
-)");
-}
-
-TEST_CASE("lvalue driver assertion regression GH #551") {
-    auto tree = SyntaxTree::fromText(R"(
-module M #(parameter int W=1) (
-    input  logic         clk,
-    input  logic         rst,
-    input  logic [W-1:0] d,
-    output logic [W-1:0] o
-);
-endmodule
-
-module M2 #(
-    parameter int W = 2
-) (
-    input logic clk,
-    input logic rst
-);
-    localparam int W_MAX = $clog2(W);
-
-    logic [W_MAX:0] d, o;
-    logic x_d, x_o;
-
-    M m [W_MAX+1:0] (
-        .clk (clk),
-        .rst (rst),
-        .d   ({x_d, d}),
-        .o   ({x_o, o})
-    );
-endmodule
-)");
-
-    Compilation compilation;
-    compilation.addSyntaxTree(tree);
-    NO_COMPILATION_ERRORS;
 }
 
 TEST_CASE("Specify path descriptions") {
